feat : 모바일/데스크탑 정보 사이드바 컴포넌트 분리 #60
